Changed files locations
